exceptions et amélioration du code

+++ src/main/java/fr/aston/sqli/projet/canadagalerie/models/nosql/Artiste.java
@@ -1,7 +1,5 @@
 package fr.aston.sqli.projet.canadagalerie.models.nosql;
 
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
 
 import lombok.AllArgsConstructor;
@@ -19,7 +17,7 @@ import lombok.ToString;
 public class Artiste {
 
 	@Field(value = "id")
-	private Integer id;
+	private Long id;
 	@Field(value = "nom")
 	private String nom;
